+++ vtapi3/__init__.py
@@ -1,5 +1,5 @@
 __title__ = 'vtapi3'
-__version__ = '1.1.3'
+__version__ = '1.2.0'
 __author__ = 'Evgeny Drobotun'
 __author_email__ = 'drobotun@xakep.ru'
 __license__ = 'MIT'